add record click item & add some small function

+++ app/src/main/java/com/cll/toy/filebrowewrlib/filebrower/managers/FileManager.kt
@@ -1,5 +1,6 @@
 package com.cll.toy.filebrowewrlib.filebrower.managers
 
+import android.util.Log
 import java.io.File
 
 /**
@@ -11,21 +12,53 @@ enum class FileManager {
     SINGLETON;
 
 
-    fun create(filePath: String) : Boolean{
-        val file = File(filePath);
-        return file.createNewFile();
+    public val TYPE_DIR : Int = 0;
+    public val TYPE_FILE : Int = 1;
+    fun create(filePath: String, type : Int) : Boolean{
+        val file = File(filePath)
+        if (type == TYPE_DIR) return file.mkdirs() else return file.createNewFile();
     }
 
-    fun create(file: File) : Boolean{
-        return file.createNewFile();
+    fun create(dir: String, file_name : String, type : Int) : Boolean{
+        val file = File(dir, file_name);
+        if (type == TYPE_DIR) return file.mkdirs() else return file.createNewFile();
+    }
+
+    fun create(file: File, type : Int) : Boolean{
+        if (type == TYPE_DIR) return file.mkdirs() else return file.createNewFile();
     }
 
     fun delete(filePath: String) : Boolean{
         val file = File(filePath);
-        return file.delete();
+        if (file.isFile){
+            return file.delete();
+        }else {
+            forFiles(file)
+            return file.delete();
+        }
     }
 
+    private fun forFiles(file: File){
+        val list = file.listFiles()
+        if (list.size > 0) {
+            for (i in 0..list.size - 1) {
+                if(list[i].isFile){
+                    delete(list[i])
+                }else{
+                    forFiles(list[i])
+                }
+            }
+        }
+    }
+
+
     fun delete(file: File) : Boolean{
+        if (!file.exists()){
+            return false;
+        }
+        if (file.isDirectory){
+            forFiles(file);
+        }
         return file.delete();
     }
 }
